// frontend/src/types.ts
import type { DataConnection } from 'peerjs'

export type UploadedFile = File & { entryFullPath?: string }


// 连接状态
export enum UploaderConnectionStatus {
    Pending = 'PENDING',
    Ready = 'READY',
    Paused = 'PAUSED',
    Uploading = 'UPLOADING',
    Done = 'DONE',
    Authenticating = 'AUTHENTICATING',
    InvalidPassword = 'INVALID_PASSWORD',
    Closed = 'CLOSED',
}

// 连接信息
export type UploaderConnection = {
    status: UploaderConnectionStatus
    dataConnection: DataConnection
    browserName?: string
    browserVersion?: string
    osName?: string
    osVersion?: string
    mobileVendor?: string
    mobileModel?: string
    uploadingFileName?: string
    uploadingOffset?: number
    completedFiles: number
    totalFiles: number
    currentFileProgress: number
}

// 文件信息
export interface FileInfo {
    // Define your FileInfo interface properties here
    // For example:
    name: string;
    size: number;
    type: string;
    // other properties...
}